requirements.txt
fastapi[all]~=0.104.1
uvicorn~=0.23.2
sqlalchemy~=2.0.23
asyncpg~=0.28.0
pydantic~=2.4.2
dadata~=21.10.1
pytest~=7.4.2
aio_pika~=9.3.0
psycopg2-binary~=2.9.9
httpx[http2]~=0.25.0
starlette~=0.27.0
alembic~=1.13.1
python-dotenv~=1.0.1